Separate vcf_writer.rs from main
